requirements.txt
attrs==19.1.0
bcrypt==3.1.7
cffi==1.12.3
falcon==2.0.0
gunicorn==19.9.0
jsonschema==3.0.1
marshmallow==2.19.5
marshmallow-sqlalchemy==0.17.0
pycparser==2.19
PyJWT==1.7.1
pyrsistent==0.15.2
python-dotenv==0.10.3
six==1.12.0
SQLAlchemy==1.3.5
